+++ InteractionNew/Runtime/Scripts/Manipulation/IManipulationSystem.cs
@@ -1,7 +1,5 @@
 using Reflectis.SDK.Core;
 
-using System.Collections.Generic;
-
 using UnityEngine;
 
 namespace Reflectis.SDK.InteractionNew
@@ -10,7 +8,8 @@ namespace Reflectis.SDK.InteractionNew
     {
         public InteractableBehaviourBase SetupInteractableBehaviour(GameObject obj);
 
-        public List<AwaitableScriptableAction> OnHoverEnterActions { get; }
-        public List<AwaitableScriptableAction> OnHoverExitActions { get; }
+        public void OnManipulableHoverEnter(IInteractable interactable);
+
+        public void OnManipulableHoverExit(IInteractable interactable);
     }
 }