+++ InteractionNew/Runtime/Scripts/VisualScripting/HoverExitEventUnit.cs
@@ -2,8 +2,10 @@ using Unity.VisualScripting;
 
 namespace Reflectis.SDK.InteractionNew
 {
-    [UnitTitle("OnInteractableHoverExit")]//Custom EventUnit to receive the event. Adding On to the unit title as an event naming convention.
-    [UnitCategory("Events\\ReflectisEvent\\GenericInteractable")]
+    [UnitTitle("GenericInteractable: HoverExit")]
+    [UnitSurtitle("Generic Interactable")]
+    [UnitShortTitle("Hover Exit")]
+    [UnitCategory("Reflectis\\Events\\GenericInteractable")]
     public class HoverExitEventUnit : GenericInteractableEventUnit
     {