+++ InteractionNew/Runtime/Scripts/VisualScripting/InteractEventUnit.cs
@@ -2,8 +2,10 @@ using Unity.VisualScripting;
 
 namespace Reflectis.SDK.InteractionNew
 {
-    [UnitTitle("OnInteractableInteract")]//Custom EventUnit to receive the event. Adding On to the unit title as an event naming convention.
-    [UnitCategory("Events\\ReflectisEvent\\GenericInteractable")]
+    [UnitTitle("GenericInteractable: Interact")]
+    [UnitSurtitle("Generic Interactable")]
+    [UnitShortTitle("Interact")]
+    [UnitCategory("Reflectis\\Events\\GenericInteractable")]
     public class InteractEventUnit : GenericInteractableEventUnit
     {